# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-офис собственника",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/processes": {
      "get": {
        "tags": [
          "processes and result reviews"
        ],
        "summary": "Processes",
        "operationId": "processes_api_v1_processes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/ProcessDefinition"
                      },
                      {
                        "$ref": "#/components/schemas/ProcessDefinitionV2"
                      },
                      {
                        "$ref": "#/components/schemas/ProcessDefinitionV3"
                      }
                    ]
                  },
                  "type": "array",
                  "title": "Response Processes Api V1 Processes Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/processes/{process_id}/runtime": {
      "get": {
        "tags": [
          "processes and result reviews"
        ],
        "summary": "Get Process Runtime",
        "description": "Autonomy ladder position and kill switches for one process.",
        "operationId": "get_process_runtime_api_v1_processes__process_id__runtime_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Process Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Process Runtime Api V1 Processes  Process Id  Runtime Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "processes and result reviews"
        ],
        "summary": "Set Process Runtime",
        "description": "Only the OWNER moves the ladder or flips a kill switch, instantly.",
        "operationId": "set_process_runtime_api_v1_processes__process_id__runtime_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Process Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRuntimeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Set Process Runtime Api V1 Processes  Process Id  Runtime Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/quality": {
      "get": {
        "tags": [
          "processes and result reviews"
        ],
        "summary": "Process Quality",
        "operationId": "process_quality_api_v1_processes_quality_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessQualityReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/controls": {
      "get": {
        "tags": [
          "processes and result reviews"
        ],
        "summary": "Process Controls",
        "description": "Summarize policy and postflight controls without returning task contents.",
        "operationId": "process_controls_api_v1_processes_controls_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessControlReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/result-review": {
      "get": {
        "tags": [
          "processes and result reviews"
        ],
        "summary": "Get Review",
        "operationId": "get_review_api_v1_tasks__task_id__result_review_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Review Api V1 Tasks  Task Id  Result Review Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "processes and result reviews"
        ],
        "summary": "Submit Review",
        "operationId": "submit_review_api_v1_tasks__task_id__result_review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultReviewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Submit Review Api V1 Tasks  Task Id  Result Review Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessControlReport": {
        "properties": {
          "window_started": {
            "type": "string",
            "format": "date-time",
            "title": "Window Started"
          },
          "office_results": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Office Results"
          },
          "postflight_completed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Postflight Completed"
          },
          "postflight_coverage": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Postflight Coverage"
          },
          "policy_decisions": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Policy Decisions"
          },
          "control_flags": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Control Flags"
          }
        },
        "type": "object",
        "required": [
          "window_started",
          "office_results",
          "postflight_completed",
          "postflight_coverage",
          "policy_decisions",
          "control_flags"
        ],
        "title": "ProcessControlReport",
        "description": "Read-only operational controls for completed office results."
      },
      "ProcessDefinition": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "result_owner_id": {
            "type": "string",
            "title": "Result Owner Id"
          },
          "reviewer_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reviewer Roles",
            "default": [
              "owner"
            ]
          },
          "acceptance_checks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Acceptance Checks"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions",
            "default": [
              "analyze",
              "prepare_payment_draft"
            ]
          },
          "forbidden_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Forbidden Actions",
            "default": [
              "real_payment"
            ]
          },
          "autonomy_level": {
            "type": "string",
            "const": "review_required",
            "title": "Autonomy Level",
            "default": "review_required"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "version",
          "title",
          "result_owner_id",
          "acceptance_checks"
        ],
        "title": "ProcessDefinition"
      },
      "ProcessDefinitionV2": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "const": 2,
            "title": "Version",
            "default": 2
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "result_owner_id": {
            "type": "string",
            "title": "Result Owner Id"
          },
          "reviewer_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reviewer Roles",
            "default": [
              "owner"
            ]
          },
          "acceptance_checks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Acceptance Checks"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions",
            "default": [
              "analyze",
              "prepare_payment_draft"
            ]
          },
          "forbidden_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Forbidden Actions",
            "default": [
              "real_payment"
            ]
          },
          "autonomy_level": {
            "type": "string",
            "const": "review_required",
            "title": "Autonomy Level",
            "default": "review_required"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/ProcessNodeSpec"
            },
            "type": "array",
            "minItems": 1,
            "title": "Nodes"
          },
          "run_budget": {
            "$ref": "#/components/schemas/RunBudget"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "result_owner_id",
          "acceptance_checks",
          "nodes",
          "run_budget"
        ],
        "title": "ProcessDefinitionV2",
        "description": "An executable, review-required process definition.\n\nIt is application configuration rather than model-generated instructions and\nis snapshotted on a task before its first node starts."
      },
      "ProcessDefinitionV3": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "const": 3,
            "title": "Version",
            "default": 3
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "result_owner_id": {
            "type": "string",
            "title": "Result Owner Id"
          },
          "reviewer_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reviewer Roles",
            "default": [
              "owner"
            ]
          },
          "acceptance_checks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Acceptance Checks"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions",
            "default": [
              "analyze",
              "prepare_payment_draft"
            ]
          },
          "forbidden_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Forbidden Actions",
            "default": [
              "real_payment"
            ]
          },
          "autonomy_level": {
            "type": "string",
            "enum": [
              "a0_shadow",
              "a1_recommendation",
              "a2_draft",
              "a3_approved_action"
            ],
            "title": "Autonomy Level",
            "default": "a2_draft"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "output_schema_version": {
            "type": "string",
            "title": "Output Schema Version"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/ProcessNodeSpec"
            },
            "type": "array",
            "minItems": 1,
            "title": "Nodes"
          },
          "run_budget": {
            "$ref": "#/components/schemas/RunBudget"
          },
          "passport": {
            "$ref": "#/components/schemas/ProcessPassport"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "result_owner_id",
          "acceptance_checks",
          "policy_version",
          "output_schema_version",
          "nodes",
          "run_budget",
          "passport"
        ],
        "title": "ProcessDefinitionV3",
        "description": "V2 executable graph + immutable V3 passport, snapshotted per run."
      },
...
      "ProcessQualityReport": {
        "properties": {
          "window_started": {
            "type": "string",
            "format": "date-time",
            "title": "Window Started"
          },
          "completed_tasks": {
            "type": "integer",
            "title": "Completed Tasks"
          },
          "reviewed_results": {
            "type": "integer",
            "title": "Reviewed Results"
          },
          "pending_reviews": {
            "type": "integer",
            "title": "Pending Reviews"
          },
          "review_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Review Rate"
          },
          "decisions": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Decisions"
          },
          "average_review_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Review Seconds"
          }
        },
        "type": "object",
        "required": [
          "window_started",
          "completed_tasks",
          "reviewed_results",
          "pending_reviews",
          "review_rate",
          "decisions"
        ],
        "title": "ProcessQualityReport"
      },
      "ProcessRuntimeUpdate": {
        "properties": {
          "autonomy_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "a0_shadow",
                  "a1_recommendation",
                  "a2_draft",
                  "a3_approved_action"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Autonomy Level"
          },
          "process_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process Enabled"
          },
          "llm_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Enabled"
          },
          "write_tools_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Write Tools Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProcessRuntimeUpdate"
      },
...
      "ResultReviewCreate": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "accepted",
              "rework_required",
              "rejected"
            ],
            "title": "Decision"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Reason"
          },
          "result_hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Result Hash"
          },
          "target_node_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Target Node Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "reason",
          "result_hash"
        ],
        "title": "ResultReviewCreate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}